@@ -13,14 +13,12 @@ class Submission < ActiveRecord::Base
   validate :must_have_valid_problem
   validate :must_specify_language
 
+  has_one :task
+
   before_save :assign_latest_number_if_new_recond
 
   def self.find_last_by_user_and_problem(user_id, problem_id)
-    last_sub = find(:first, 
-                    :conditions => {:user_id => user_id,
-                      :problem_id => problem_id},
-                    :order => 'number DESC')
-    return last_sub
+    where("user_id = ? AND problem_id = ?",user_id,problem_id).last
   end
 
   def self.find_all_last_by_problem(problem_id)
@@ -43,7 +41,7 @@ class Submission < ActiveRecord::Base
 
   def self.find_last_for_all_available_problems(user_id)
     submissions = Array.new
-    problems = Problem.find_available_problems
+    problems = Problem.available_problems
     problems.each do |problem|
       sub = Submission.find_last_by_user_and_problem(user_id, problem.id)
       submissions << sub if sub!=nil
@@ -52,20 +50,11 @@ class Submission < ActiveRecord::Base
   end
 
   def self.find_by_user_problem_number(user_id, problem_id, number)
-    Submission.find(:first,
-                    :conditions => {
-                      :user_id => user_id,
-                      :problem_id => problem_id,
-                      :number => number
-                    })
+    where("user_id = ? AND problem_id = ? AND number = ?",user_id,problem_id,number).first
   end
 
   def self.find_all_by_user_problem(user_id, problem_id)
-    Submission.find(:all,
-                    :conditions => {
-                      :user_id => user_id,
-                      :problem_id => problem_id,
-                    })
+    where("user_id = ? AND problem_id = ?",user_id,problem_id)
   end
 
   def download_filename
